Inherit `Layers` from `Array`
Needed `#size`, and realized this class basically encapsulates a lone
`Array` anyway. Only caveat to this is that we cannot cache the layers
by name anymore, because `Array` is implemented as a C struct. I tried
to benchmark the performance impact of this, but it was so neglibible
that the variance from run to run (with 1e500 layers) was greater than
the difference between caching and not caching.